moved eventlisteners to their own file
