+++ public/js/script.js
@@ -1,32 +1,9 @@
-import { API } from './api.js';
-import { Generator } from './htmlGen.js';
+import { EventListeners } from './event.js'
 
 function qS(selector) {
   return document.querySelector(selector);
 }
 
-qS('header input').addEventListener('keydown', async (e) => {
-  if(e.code != 'Enter') return;
-  qS('main').classList.add('triple');
-  qS('main').classList.remove('double');
-  const value = qS('header input').value;
-  Generator.clear();
-  const repos = await API.search(value);
-  repos.forEach(repo => {
-      Generator.repoCard(repo);
-  });
-})
+qS('header input').addEventListener('keydown', () => {EventListeners.search(event)})
 
-qS('main').addEventListener('click', async (e)  => {
-  if(e.target.classList.contains('forks_url')) {
-    e.preventDefault();
-    qS('main').classList.remove('triple');
-    qS('main').classList.add('double');
-    const href = e.target.href;
-    Generator.clear();
-    const forks = await API.forks(href);
-    await forks.forEach(async fork => {
-      await Generator.forkCard(fork);
-    })
-  }
-})
+qS('main').addEventListener('click', ()  => {EventListeners.forks(event)})
